Updated Search to return full details of recipes by default. Made draft Summary Screen for each Recipe, though incomplete. Fixed minor issues in the data classes to accommodate the true range of response types (Doubles instead of Ints). Prepared a Draft TestRecipesPanel in case I decide to switch how I present the Recipe Cards.

+++ app/src/main/java/com/brentcodes/recipesapplication/model/spoonaculardata/Flavonoids.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class Flavonoids(
 
     @SerializedName("name") var name: String? = null,
-    @SerializedName("amount") var amount: Int? = null,
+    @SerializedName("amount") var amount: Double? = null,
     @SerializedName("unit") var unit: String? = null
 
 )
