+++ app/src/main/java/com/brentcodes/recipesapplication/model/spoonaculardata/Nutrients.kt
@@ -8,6 +8,6 @@ data class Nutrients(
     @SerializedName("name") var name: String? = null,
     @SerializedName("amount") var amount: Double? = null,
     @SerializedName("unit") var unit: String? = null,
-    @SerializedName("percentOfDailyNeeds") var percentOfDailyNeeds: Int? = null
+    @SerializedName("percentOfDailyNeeds") var percentOfDailyNeeds: Double? = null
 
 )
